requirements.txt
absl-py==2.3.1
accelerate==1.10.1
affine==2.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
attrs==25.4.0
black==25.11.0
certifi==2020.6.20
click==8.1.8
click-plugins==1.1.1.2
cligj==0.7.2
colored==2.3.1
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
eval_type_backport==0.3.1
exceptiongroup==1.3.1
flake8==7.3.0
flatbuffers==25.12.19
fonttools==4.60.2
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
GPUtil==1.4.0
grpcio==1.78.0
grpcio-tools==1.78.0
h5py==3.14.0
hf-xet==1.2.0
huggingface_hub==0.36.2
humanfriendly==10.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.1.0
joblib==1.5.3
kiwisolver==1.4.7
Markdown==3.9
markdown-it-py==3.0.0
matplotlib==3.9.4
mccabe==0.7.0
mdurl==0.1.2
mkl-service==2.4.0
ml_dtypes==0.5.4
mypy_extensions==1.1.0
numpy==1.26.4
nvidia-ml-py==13.590.48
onnx==1.19.1
onnx-graphsurgeon==0.5.2
onnxruntime==1.19.2
onnxsim==0.4.36
opencv-python==4.5.4.60
opencv-python-headless==4.13.0.92
pandas==2.3.3
pathspec==1.0.4
peft==0.17.1
pluggy==1.6.0
polygraphy==0.49.26
protobuf==6.33.5
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
pynvml==13.0.1
pyparsing==3.3.2
pytest==8.4.2
pytokens==0.4.1
pytz==2025.2
rasterio==1.4.3
regex==2026.1.15
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.53.0
simsimd==6.5.13
smmap==5.0.2
stringzilla==4.6.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
timm==1.0.24
tokenizers==0.22.2
tomli==2.4.0
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.67.3
transformers==4.57.6
monai==1.3.2
fastapi==0.115.12
uvicorn==0.30.6
prometheus-client==0.21.0
typing-inspection==0.4.2
tzdata==2025.3
wandb==0.25.0
wcwidth==0.1.9
Werkzeug==3.1.5
zipp==3.23.0
